change parameters in physics components to qreal from int

+++ include/physicscomponent.h
@@ -12,7 +12,7 @@ protected:
     int newx,newy;
     std::vector<std::vector<Tile*> > tilesMap;
     QGraphicsScene * scene;
-    int width_of_tile,height_of_tile,screenWidth,screenHeight;
+    qreal width_of_tile,height_of_tile,screenWidth,screenHeight;
     int curJumpCount,maxJumpCount;
 public:
     PhysicsComponent() {};
